backend/requirements/base.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
starlette>=0.40.0
jinja2>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
httpx>=0.25.0
redis>=5.0.0
pika>=1.3.0
motor>=3.3.0
pymongo[srv]>=4.6.0
prometheus-client>=0.17.0
websockets>=11.0.0
aiofiles>=23.0.0
python-multipart>=0.0.6
types-redis
types-pika
mypy 